+++ builder/commands/optioncmd.py
@@ -36,12 +36,13 @@ class OptionParser(object):
         NOTE:
             -p, --plot: output a plot.
             -r, --rubi: output with rubi.
-            --comment: output with comments.
+            -c, --comment: output with comments.
             --console: output to a console.
             --debug: set DEBUG on log leve.
             --forcemecab: specific using travic-ci
             --format: format type for output
             --log: set log level.
+            --part: select story parts.
             --priority: set a priority filter.
             --text: output as a text.
         Returns:
@@ -52,14 +53,15 @@ class OptionParser(object):
         test_data = test_data if test_data else []
 
         LOG.info('OPT_PARSE: set option arguments')
+        parser.add_argument('-c', '--comment', help='output with comments', action='store_true')
         parser.add_argument('-p', '--plot', help='output a plot', action='store_true')
         parser.add_argument('-r', '--rubi', help='output with rubi', action='store_true')
-        parser.add_argument('--comment', help='output with comments', action='store_true')
         parser.add_argument('--console', help='output to the console (for debug)', action='store_true')
         parser.add_argument('--debug', help='set DEBUG on log level', action='store_true')
         parser.add_argument('--forcemecab', help='force no use mecab dir', action='store_true')
         parser.add_argument('--format', help='format type for output', type=str)
         parser.add_argument('--log', help='set a log level', type=str)
+        parser.add_argument('--part', help='select story parts', type=str)
         parser.add_argument('--priority', help='set a priority filter', type=int)
         parser.add_argument('--text', help='output as a text', action='store_true')